requirements.txt
aiohttp==3.6.2
asn1crypto==0.24.0
async-timeout==3.0.1
attrs==19.3.0
Automat==0.6.0
certifi==2018.1.18
cffi==1.14.0
chardet==3.0.4
click==6.7
colorama==0.3.7
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
cryptography==2.1.4
discord.py==1.3.3
distro-info===0.18ubuntu0.18.04.1
docopt==0.6.2
httplib2==0.9.2
hyperlink==17.3.1
idna==2.9
idna-ssl==1.1.0
incremental==16.10.1
keyring==10.6.0
keyrings.alt==3.0
language-selector==0.1
multidict==4.7.6
netifaces==0.10.4
PAM==0.4.2
pipreqs==0.4.10
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycparser==2.20
pycrypto==2.6.1
pygobject==3.26.1
pymongo==3.10.1
PyNaCl==1.3.0
pyOpenSSL==17.5.0
pyserial==3.4
python-apt==1.6.5+ubuntu0.2
python-debian==0.1.32
python-dotenv==0.13.0
pyxattr==0.6.0
pyxdg==0.25
PyYAML==3.12
requests==2.18.4
requests-unixsocket==0.1.5
riotwatcher==3.0.0
SecretStorage==2.3.1
service-identity==16.0.0
six==1.15.0
ssh-import-id==5.7
systemd-python==234
Twisted==17.9.0
typing-extensions==3.7.4.2
ufw==0.36
unattended-upgrades==0.1
urllib3==1.22
websockets==8.1
yarg==0.1.9
yarl==1.4.2
youtube-dl==2020.5.8
zope.interface==4.3.2
